refactor(scaffold): move the template guards' readers into tests
`pinnedCliVersion`, `pinIsAhead`, and `declaredRunName` had no runtime
caller: nothing reads a scaffolded template's *contents* at runtime —
`init` copies the files through byte for byte, and
scripts/sync-workflow-version.mjs deliberately keeps its own copy of the
pin pattern to stay a plain node script with no build step. Parsing a
template is only ever something an assertion wants, which showed up as
tests/core/scaffold.test.ts unit-testing parsers that only tests call.

They move to tests/helpers/workflow-template.ts, shared by the two
scaffold suites that hold the templates to their contracts. core/scaffold
keeps what `init` actually calls.

`declaredRunName` gains the YAML semantics it was missing: a plain scalar
ends at the first ` #`, so an unquoted run-name silently truncates to
"border-collie worker" — which the reader, echoing the line as written,
would have waved through, guarding a contract against exactly the mistake
it could not see. It now reads the value the way YAML does and the guard
fails on it, verified against a deliberately unquoted template.
